Move ProteinSearchViewProvider to Protein. Stop special casing (and double rendering) specific views in DoProteinSearchAction.
