+++ protein/src/org/labkey/protein/ProteinController.java
@@ -18,6 +18,7 @@ package org.labkey.protein;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.apache.commons.lang3.StringUtils;
+import org.jetbrains.annotations.Nullable;
 import org.labkey.api.action.FormHandlerAction;
 import org.labkey.api.action.FormViewAction;
 import org.labkey.api.action.LabKeyError;
@@ -25,6 +26,7 @@ import org.labkey.api.action.QueryViewAction;
 import org.labkey.api.action.SimpleViewAction;
 import org.labkey.api.action.SpringActionController;
 import org.labkey.api.admin.AdminUrls;
+import org.labkey.api.annotations.Migrate;
 import org.labkey.api.collections.CaseInsensitiveHashSet;
 import org.labkey.api.data.ActionButton;
 import org.labkey.api.data.ButtonBar;
@@ -41,37 +43,41 @@ import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.Sort;
 import org.labkey.api.data.SqlSelector;
 import org.labkey.api.data.Table;
+import org.labkey.api.data.TableInfo;
 import org.labkey.api.exp.DomainDescriptor;
 import org.labkey.api.exp.Lsid;
 import org.labkey.api.exp.OntologyManager;
 import org.labkey.api.exp.PropertyDescriptor;
 import org.labkey.api.exp.PropertyType;
-import org.labkey.api.module.ModuleLoader;
 import org.labkey.api.pipeline.PipelineService;
+import org.labkey.api.protein.MatchCriteria;
+import org.labkey.api.protein.ProteinManager;
+import org.labkey.api.protein.ProteinSchema;
+import org.labkey.api.protein.ProteinService;
 import org.labkey.api.protein.annotation.AnnotationInsertion;
 import org.labkey.api.protein.annotation.CustomAnnotationImportHelper;
 import org.labkey.api.protein.annotation.CustomAnnotationSet;
 import org.labkey.api.protein.annotation.CustomAnnotationSetManager;
 import org.labkey.api.protein.annotation.CustomAnnotationType;
 import org.labkey.api.protein.annotation.DefaultAnnotationLoader;
-import org.labkey.api.protein.search.PepSearchModel;
-import org.labkey.api.protein.search.PeptideFilterSearchForm;
-import org.labkey.api.protein.search.PeptideSearchForm;
 import org.labkey.api.protein.annotation.ProtSprotOrgMap;
 import org.labkey.api.protein.annotation.ProteinAnnotationPipelineProvider;
-import org.labkey.api.protein.ProteinManager;
-import org.labkey.api.protein.ProteinSchema;
-import org.labkey.api.protein.ProteinService;
 import org.labkey.api.protein.annotation.XMLProteinLoader;
 import org.labkey.api.protein.fasta.FastaDbLoader;
 import org.labkey.api.protein.fasta.FastaParsingForm;
 import org.labkey.api.protein.fasta.FastaReloaderJob;
 import org.labkey.api.protein.go.GoLoader;
 import org.labkey.api.protein.query.CustomAnnotationSchema;
+import org.labkey.api.protein.query.ProteinUserSchema;
+import org.labkey.api.protein.query.SequencesTableInfo;
+import org.labkey.api.protein.search.PepSearchModel;
+import org.labkey.api.protein.search.PeptideFilterSearchForm;
+import org.labkey.api.protein.search.PeptideSearchForm;
 import org.labkey.api.protein.search.ProbabilityProteinSearchForm;
 import org.labkey.api.protein.search.ProteinSearchForm;
 import org.labkey.api.protein.search.ProteinSearchWebPart;
 import org.labkey.api.query.FieldKey;
+import org.labkey.api.query.QueryService;
 import org.labkey.api.query.QuerySettings;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.QueryViewProvider;
@@ -80,6 +86,7 @@ import org.labkey.api.query.ValidationError;
 import org.labkey.api.query.ValidationException;
 import org.labkey.api.reader.ColumnDescriptor;
 import org.labkey.api.reader.TabLoader;
+import org.labkey.api.reports.ReportService;
 import org.labkey.api.security.AdminConsoleAction;
 import org.labkey.api.security.RequiresPermission;
 import org.labkey.api.security.RequiresSiteAdmin;
@@ -108,6 +115,7 @@ import org.labkey.api.view.NotFoundException;
 import org.labkey.api.view.RedirectException;
 import org.labkey.api.view.TabStripView;
 import org.labkey.api.view.VBox;
+import org.labkey.api.view.ViewContext;
 import org.labkey.api.view.WebPartView;
 import org.labkey.api.view.template.PageConfig;
 import org.springframework.validation.BindException;
@@ -237,29 +245,13 @@ public class ProteinController extends SpringActionController
 
             VBox result = new VBox(searchFormView);
 
-            if (form.isShowMatchingProteins())
-            {
-                QueryView proteinsView = createInitializedQueryView(form, errors, false, ProteinSearchForm.POTENTIAL_PROTEIN_DATA_REGION);
-                proteinsView.enableExpandCollapse("ProteinSearchProteinMatches", true);
-
-                result.addView(proteinsView);
-            }
-
-            if (form.isShowProteinGroups() &&
-                    // Add the "Protein Group Results" web part only if the targetedms module is not enabled in the container.
-                    !getContainer().getActiveModules().contains(ModuleLoader.getInstance().getModule("targetedms")))
-            {
-                QueryView groupsView = createInitializedQueryView(form, errors, false, ProteinSearchForm.PROTEIN_DATA_REGION);
-                groupsView.enableExpandCollapse("ProteinSearchGroupMatches", false);
-                result.addView(groupsView);
-            }
-
             for (QueryViewProvider<ProteinSearchForm> provider : ProteinService.get().getProteinSearchViewProviders())
             {
                 QueryView queryView = provider.createView(getViewContext(), form, errors);
                 if (queryView != null)
                     result.addView(queryView);
             }
+
             return result;
         }
 
@@ -268,6 +260,67 @@ public class ProteinController extends SpringActionController
         {
             setHelpTopic("proteinSearch");
             root.addChild("Protein Search Results");
+        }
+    }
+
+    public static class ProteinSearchViewProvider implements QueryViewProvider<ProteinSearchForm>
+    {
+        @Override
+        public String getDataRegionName()
+        {
+            return ProteinSearchForm.POTENTIAL_PROTEIN_DATA_REGION;
+        }
+
+        @Migrate // address getContainerCondition()
+        @Override
+        public @Nullable QueryView createView(ViewContext ctx, ProteinSearchForm form, BindException errors)
+        {
+            QueryView proteinsView = null;
+
+            if (form.isShowMatchingProteins())
+            {
+                UserSchema schema = QueryService.get().getUserSchema(ctx.getUser(), ctx.getContainer(), ProteinUserSchema.NAME);
+
+                if (null != schema)
+                {
+
+                    QuerySettings proteinsSettings = schema.getSettings(ctx, ProteinSearchForm.POTENTIAL_PROTEIN_DATA_REGION);
+                    proteinsSettings.setQueryName(ProteinUserSchema.TableType.Sequences.toString());
+                    proteinsView = new QueryView(schema, proteinsSettings, errors)
+                    {
+                        @Override
+                        protected TableInfo createTable()
+                        {
+                            ProteinUserSchema schema = (ProteinUserSchema) getSchema();
+                            return schema.createSequences();
+                        }
+                    };
+                    // Disable R and other reporting until there's an implementation that respects the search criteria
+                    proteinsView.setViewItemFilter(ReportService.EMPTY_ITEM_LIST);
+
+                    proteinsView.setButtonBarPosition(DataRegion.ButtonBarPosition.TOP);
+                    SequencesTableInfo<ProteinUserSchema> sequencesTableInfo = (SequencesTableInfo<ProteinUserSchema>) proteinsView.getTable();
+                    // TODO!!
+//            ((ProbabilityProteinSearchForm)form).setRestrictCondition(MS2Controller.getContainerCondition(ctx.getContainer(), ctx.getUser()));
+                    int[] seqIds = form.getSeqId();
+                    if (seqIds.length <= 500)
+                    {
+                        sequencesTableInfo.addSeqIdFilter(seqIds);
+                    }
+                    else
+                    {
+                        sequencesTableInfo.addProteinNameFilter(form.getIdentifier(), form.isExactMatch() ? MatchCriteria.EXACT : MatchCriteria.PREFIX);
+                        if (form.isRestrictProteins())
+                        {
+                            // TODO!!
+//                    sequencesTableInfo.addCondition(getContainerCondition(ctx.getContainer(), ctx.getUser()));
+                        }
+                    }
+                    proteinsView.setTitle("Matching Proteins (" + (seqIds.length == 0 ? "None" : seqIds.length) + ")");
+                }
+            }
+
+            return proteinsView;
         }
     }
